added styling file
